SPARKC-635 fix session counting tests
Some of the session tests failed because of an incorrect number of cached
sessions.
The problem was revealed by a new test that created sessions with a test
connection factory. The sessions were populated scanner in
CassandraScanPartitionReaderFactory. The scanner was reinitialized couple
of time during the test.
This commit introduces scanner "caching" (val instead of def) and thus
removed the reinitialization. Also, now the scanner is properly
closed.